Fix IndexOutOfBound exception that occurs at configuration
If you apply this plugin in a fresh develop branch where there are no past commits, an IndexOutOfBounds exception occurs during configuration. This is because the plugin tries to figure out the alpha version by adding 1 to the number of commits, but the operation leads to a -1 as the index.